refactor: 🎨 提取公共定义和常量
