+++ packages/preset-umi/src/libs/folderCache/constant.ts
@@ -1,0 +1,13 @@
+export const DEFAULT_SRC_IGNORES = [
+  '**/*.d.ts',
+  '**/*.{test,spec}.{js,ts,jsx,tsx}',
+  '**/cypress/**',
+  '**/.umi-production/**',
+  '**/.umi-test/**',
+  '**/node_modules/**',
+  '**/.git/**',
+  '**/dist/**',
+  '**/coverage/**',
+  '**/jest.config.{ts,js}',
+  '**/jest-setup.{ts,js}',
+];
