+++ packages/preset-umi/src/libs/folderCache/types.ts
@@ -1,0 +1,11 @@
+type FileContent = string;
+
+export type AbsPath = string;
+export type FileContentCache = Record<AbsPath, FileContent>;
+
+export type FileChangeEvent = {
+  event: 'unlink' | 'change' | 'add';
+  path: string;
+};
+
+export type Event = FileChangeEvent['event'];
